# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateOrderInput:
    properties:
      addressId:
        type: integer
      deliveryTime:
        type: string
      items:
        items:
          $ref: '#/definitions/domain.CreateOrderItemInput'
        type: array
      totalPrice:
        type: number
    type: object
  domain.CreateOrderItemInput:
    properties:
      price:
        type: number
      productId:
        type: integer
      quantity:
        type: integer
    type: object
  domain.Order:
    properties:
      addressId:
        type: integer
      createdAt:
        type: string
      deliveryTime:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/domain.OrderItem'
        type: array
      status:
        type: string
      totalPrice:
        type: number
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  domain.OrderItem:
    properties:
      id:
        type: integer
      orderId:
        type: integer
      price:
        type: number
      productId:
        type: integer
      quantity:
        type: integer
    type: object
  domain.UpdateStatusInput:
    properties:
      status:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.MessageResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8083
info:
  contact: {}
  description: API for order management and processing
  title: Mixfood Order Service API
  version: "1.0"
paths:
  /api/admin/orders:
    get:
      parameters:
      - description: Limit of records
        in: query
        name: limit
        type: integer
      - description: Offset of records
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all orders (Admin only)
  /api/admin/orders/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateStatusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update order status (Admin only)
  /api/orders:
    get:
      parameters:
      - description: Limit of records
        in: query
        name: limit
        type: integer
      - description: Offset of records
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user's orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Order details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.CreateOrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new order
securityDefinitions:
  BearerAuth:
    description: 'Enter token in format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"